requirements.txt
flask
numpy
pandas
joblib
mysql-connector-python
scikit-learn
torch
nltk
